requirements.txt
fastapi>=0.110,<0.115
uvicorn[standard]>=0.27,<0.31
aiogram>=3.4,<3.14
SQLAlchemy>=2.0,<2.1
asyncpg>=0.29,<0.30
alembic>=1.13,<1.14
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-dotenv>=1.0
httpx>=0.27
Pillow>=10.2
qrcode>=7.4
openpyxl>=3.1
